refac(compiler): flatten crate structure by moving out of generators subcrate, resolve clippy warnings
