--- src/generators/mod.rs
@@ -1,2 +1,0 @@
-pub(crate) mod mermaid;
-use mermaid::*;
